feat(@argdown/core): changes and fixes regarding groups.
Added ClosedGroupPlugin, groups can now be closed.
The ClosedGroupPlugin will remove the child nodes of closed groups
and transform their relations to group relations.

Added group.sections config options to select which sections
will be used as groups and which groups will be closed.

Added group.ignoreIsClosed and group.ignoreIsGroup config options to ignore data flags.

Fixed groupLevel and sectionLevel assignment.

Fixed section assignment for nodes that are in sections that are no groups, but have ancestor sections that are.

Changed section assignments for equivalent classes without definitions and arguments without descriptions or pcs. In these cases the section of the first reference is now assigned.
closes #65